requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==40.0.2
decouple==0.0.7
defusedxml==0.7.1
Django==4.1.6
django-allauth==0.54.0
django-crispy-forms==2.0
django-filter==23.1
django-mathfilters==1.0.0
gunicorn==20.1.0
idna==3.4
oauthlib==3.2.2
Pillow==9.4.0
pycparser==2.21
PyJWT==2.6.0
python3-openid==3.2.0
requests==2.28.2
requests-oauthlib==1.3.1
six==1.15.0
sqlparse==0.4.3
urllib3==1.26.14
